Change variable typ

+++ client/Assets/Scripts/Skills/SkillInfo.cs
@@ -12,5 +12,5 @@ public class SkillInfo : ScriptableObject
     public float animationSpeedMultiplier;
     public bool hasModelAnimation;
     public AudioClip abilityStartSfx;
-    public float skillCircleRadius;
+    public ulong skillCircleRadius;
 }
